make supre npm env prepared before start mcp client
